+++ .github/workflows/draft-pdf.yml
@@ -20,7 +20,7 @@ jobs:
         with:
           journal: joss
           # This should be the path to the paper within your repo.
-          paper-path: joss/paper.md
+          paper-path: joss_paper/paper.md
       - name: Upload
         uses: actions/upload-artifact@v4
         with:
@@ -28,4 +28,4 @@ jobs:
           # This is the output path where Pandoc will write the compiled
           # PDF. Note, this should be the same directory as the input
           # paper.md
-          path: joss/paper.pdf
+          path: joss_paper/paper.pdf
