### 2025-06-04 #8338 @kthyng
@karlropkins Thanks for fixing that!
* Do you have tests?
* I see the descriptive preprint, but how about some docs like with tutorials/examples, etc?
* It would be great to shorten the paper to closer to 1000 words if possible
