+++ joss_paper/paper.md
@@ -47,7 +47,7 @@ environments.
 # Statement of Need
 
 Authorities responsible for air quality management are typically required to implement and 
-evaluate the air quality interventions they adopt [@bradley2019review]. These interventions 
+evaluate air quality interventions they adopt [@bradley2019review]. These interventions 
 are often costly, disruptive and unpopular [@glazener2019transforming], and associated 
 benefits are often hard to measure because of the inherent variability of air quality data 
 [@kelly2011impact; @pearce2011quantifying; @jones2012large; @grange2019using]. 
@@ -59,7 +59,7 @@ been developed for their use, e.g. `bcp` [@erdman2008bcp], `changepoint` [@killi
 been applied to the detection of large changes in air quality time-series [@carslaw2006change], 
 and with additional signal isolation to smaller changes [@carslaw2007detecting]. However, 
 those tasked with air quality data analysis, although highly skilled in a wide range of air 
-quality monitoring and assessment activities, are unlike to be able to dedicate sufficient 
+quality assessment activities, are unlike to be able to dedicate sufficient 
 time and funding to the development of in-house expertise required to routinely apply these. 
 
 `AQEval` was developed to address this skill gap. It aligns the inputs and outputs of a number 
@@ -69,92 +69,88 @@ routinely detect, characterise and quantify discrete change in air quality data.
 As many air quality professionals already use the `R` package `openair` 
 [@carslaw2012openair; @ropkins2012openair] for more conventional analysis and data visualisation, 
 `AQEval` has also been written using `openair` coding conventions and function structure to reduce 
-the learning-curve typical associated with learning new software. 
+the learning-curve typically associated with learning new software. 
 
 
 # Sources
 
-`AQEval` is freely available under General Public License (GPL) from conventional on-line R archives: 
+`AQEval` is freely available under General Public License (GPL) from conventional R archives: 
 
 -	The latest (stable) release version of `AQEval` is on the Comprehensive R Archive Network (CRAN) at 
-[https://CRAN.R-project.org/package=AQEval](https://CRAN.R-project.org/package=AQEval); and, 
+[https://CRAN.R-project.org/package=AQEval](https://CRAN.R-project.org/package=AQEval);  
 -	The developers’ version and code are publicly available on GitHub at 
 [https://github.com/karlropkins/AQEval](https://github.com/karlropkins/AQEval), which can also be 
-used to report issues or suggest changes. 
+used to report issues or suggest changes; and 
 -	The project website is at [https://karlropkins.github.io/AQEval/](https://karlropkins.github.io/AQEval/).
 
 
 # Analytical Rationale
 
-The `AQEval` Break-Point/Segment (BP/S) methods involve three steps: finding possible ‘points of change’, 
-using break-points, and then characterising and quantifying ‘regions of change’ about these as segments:  
+The `AQEval` Break-Point/Segment (BP/S) methods involve three steps: finding possible ‘points-of-change’, 
+using break-points, and then characterising and quantifying ‘regions-of-change’ for these using segments:  
 
 1. Breaks-points are found using the `strucchange` methods of Zeileis and colleagues 
    [@zeileis2002strucchange; @zeileis2003testing]. Here, a rolling-window approach is applied: 
    a first subset of data (or time-series window, TW~0~ in \autoref{fig:1}a), is selected and a linear 
-   regression model built; the window is then rolled on, typically one measurement (TW~1~ in 
+   regression model built; the window is then advanced, typically one measurement (TW~1~ in 
    \autoref{fig:1}a) and a second model build; and so on, until the end of the time-series; then, 
[…]
