+++ .github/workflows/draft-pdf.yml
@@ -1,6 +1,6 @@
 name: Draft PDF
 on:
-  push:
+  workflow_dispatch:
     paths:
       - joss_paper/paper.md
       - joss_paper/paper.bib
