+++ joss_paper/paper.md
@@ -36,7 +36,7 @@ bibliography: paper.bib
 
 # Summary
 
-`AQEval` (Air Quality Evaluation) is an `R` package of functions for the routine investigation 
+`AQEval` (Air Quality Evaluation) is an `R` package for the routine investigation 
 of discrete changes in air quality time-series, e.g. assessing the impact of interventions 
 and/or pollution events. The main functions use break-point/segmentation (BP/S) methods to 
 detect, characterise and quantify change, while other functions build on these to provide 
@@ -53,18 +53,18 @@ benefits are often hard to measure because of the inherent variability of air qu
 [@kelly2011impact; @pearce2011quantifying; @jones2012large; @grange2019using]. 
 
 Various methods have been developed to investigate discrete changes in a wide range of time-series 
-[see e.g. @reeves2007review; @truong2020selective], and several `R` [@r2025r] packages have 
+[see e.g. @reeves2007review; @truong2020selective]. Several `R` [@r2025r] packages have 
 been developed for their use, e.g. `bcp` [@erdman2008bcp], `changepoint` [@killick2016changepoint], 
-`segmented` [@muggeo2008segmented], and `strucchange` [@zeileis2002strucchange]. Some have even 
-been applied to the detection of large changes in air quality time-series [@carslaw2006change], 
-and with additional signal isolation to smaller changes [@carslaw2007detecting]. However, 
-those tasked with air quality data analysis, although highly skilled in a wide range of air 
-quality assessment activities, are unlike to be able to dedicate sufficient 
-time and funding to the development of in-house expertise required to routinely apply these. 
+`segmented` [@muggeo2008segmented], and `strucchange` [@zeileis2002strucchange], and some have even 
+been applied to air quality time-series, see e.g. [@carslaw2006change, @carslaw2007detecting]. 
+
+However, many of those tasked with air quality policy assessment, although highly skilled in a 
+wide range of monitoring activities, are unlike to be able to dedicate sufficient time and resources 
+to the development of in-house expertise in such specialist procedures. 
 
 `AQEval` was developed to address this skill gap. It aligns the inputs and outputs of a number 
 of statistical methods to provide a one-package option for anyone interested in using `R` to 
-routinely detect, characterise and quantify discrete change in air quality data. 
+routinely investigate change in air quality data. 
 
 As many air quality professionals already use the `R` package `openair` 
 [@carslaw2012openair; @ropkins2012openair] for more conventional analysis and data visualisation, 
@@ -74,76 +74,75 @@ the learning-curve typically associated with learning new software.
 
 # Sources
 
-`AQEval` is freely available under General Public License (GPL) from conventional R archives: 
+`AQEval` is freely available under General Public License (GPL): 
 
--	The latest (stable) release version of `AQEval` is on the Comprehensive R Archive Network (CRAN) at 
+-	The latest (stable) release version of `AQEval` is on the Comprehensive R Archive Network (CRAN) 
 [https://CRAN.R-project.org/package=AQEval](https://CRAN.R-project.org/package=AQEval);  
--	The developers’ version and code are publicly available on GitHub at 
-[https://github.com/karlropkins/AQEval](https://github.com/karlropkins/AQEval), which can also be 
-used to report issues or suggest changes; and 
+-	The developers’ version and code are publicly on GitHub 
+[https://github.com/karlropkins/AQEval](https://github.com/karlropkins/AQEval), where issues or change 
+requests can also be posted; and 
 -	The project website is at [https://karlropkins.github.io/AQEval/](https://karlropkins.github.io/AQEval/).
 
 
 # Analytical Rationale
 
[…]
