+++ joss_paper/paper.md
@@ -37,38 +37,33 @@ bibliography: paper.bib
 # Summary
 
 `AQEval` (Air Quality Evaluation) is an `R` package for the routine investigation 
-of discrete changes in air quality time-series, e.g. assessing the impact of interventions 
-and/or pollution events. The main functions use break-point/segmentation (BP/S) methods to 
-detect, characterise and quantify change, while other functions build on these to provide 
-a workflow for those looking to measure smaller changes and/or changes in more complex 
-environments. 
+of discrete changes in air quality time-series. The main functions use 
+break-point/segmentation (BP/S) methods to detect, characterise and quantify change, 
+while other functions build on these to provide a workflow for those looking to 
+measure smaller changes and/or changes in more complex environments. 
 
 
 # Statement of Need
 
 Authorities responsible for air quality management are typically required to implement and 
 evaluate air quality interventions they adopt [@bradley2019review]. These interventions 
-are often costly, disruptive and unpopular [@glazener2019transforming], and associated 
-benefits are often hard to measure because of the inherent variability of air quality data 
-[@kelly2011impact; @pearce2011quantifying; @jones2012large; @grange2019using]. 
-
-Various methods have been developed to investigate discrete changes in a wide range of time-series 
-[see e.g. @reeves2007review; @truong2020selective]. Several `R` [@r2025r] packages have 
+are often costly, disruptive and unpopular [@glazener2019transforming], and the inherent 
+variability monitoring data hinders impact assessments 
+[@kelly2011impact; @pearce2011quantifying; @jones2012large; @grange2019using]. Various methods 
+have been developed to investigate discrete changes in a wide range of time-series 
+[see e.g. @reeves2007review; @truong2020selective] and several `R` [@r2025r] packages have 
 been developed for their use, e.g. `bcp` [@erdman2008bcp], `changepoint` [@killick2016changepoint], 
-`segmented` [@muggeo2008segmented], and `strucchange` [@zeileis2002strucchange], and some have even 
+`segmented` [@muggeo2008segmented], and `strucchange` [@zeileis2002strucchange]. Some have even 
 been applied to air quality time-series, see e.g. [@carslaw2006change, @carslaw2007detecting]. 
-
 However, many of those tasked with air quality policy assessment, although highly skilled in a 
 wide range of monitoring activities, are unlike to be able to dedicate sufficient time and resources 
-to the development of in-house expertise in such specialist procedures. 
-
+to the development of in-house expertise in such specialist analyses. 
 `AQEval` was developed to address this skill gap. It aligns the inputs and outputs of a number 
 of statistical methods to provide a one-package option for anyone interested in using `R` to 
 routinely investigate change in air quality data. 
-
 As many air quality professionals already use the `R` package `openair` 
 [@carslaw2012openair; @ropkins2012openair] for more conventional analysis and data visualisation, 
-`AQEval` has also been written using `openair` coding conventions and function structure to reduce 
+`AQEval` has also been written using `openair` coding conventions to reduce 
 the learning-curve typically associated with learning new software. 
 
 
@@ -109,21 +104,21 @@ testing these and quantifying ‘regions-of-change’ about the most likely:
 trafficked site in central London where a change event (*ca.* 25 $\mu$g.m^-3^; 31%) is detected between 
 2003-01-11 and 2003-02-19. 
 
[…]
