### 2025-06-05 #8338 @karlropkins
@kthyng

I've got it down a couple of hundred words, 1,343 to 1,156 words, but running out of easy targets... That do you?

There are some examples in extended introduction (https://karlropkins.github.io/AQEval/articles/AQEval_Intro_Preprint.pdf) which we link to in the 'other outputs' section and in the package help documentation.

Guessing you are after something more?

Likewise, the only tests I am routinely running are R package builds. The checking runs all the examples in the packages documentation and that is all good. But, again, are you after more?

If so, could you point me in direction of a couple of examples of 'best practices' and I'll have a think about what we can do?
