### 2025-06-24 #8338 @kthyng
@karlropkins Sorry for my delay.

> I've got it down a couple of hundred words, 1,343 to 1,156 words, but running out of easy targets... That do you?

Yes that is fine.

In terms of examples and tests as well as other review criteria, this is a useful guide: https://joss.readthedocs.io/en/latest/review_criteria.html
It shows a range of possibilities for review items from not-acceptable to ok to good. Can you take a look at that and see where your submission fits?
