+++ tests/tinytest.R
@@ -1,0 +1,5 @@
+
+if ( requireNamespace("tinytest", quietly=TRUE) ){
+  tinytest::test_package("AQEval")
+}
+
