+++ NEWS.md
@@ -4,6 +4,7 @@
     * __v 0.6.0__ [2024-12-19] added data.table method to calcDateRangeStat 
     * __v 0.6.0__ [2025-01-09] CRAN update 
     * __v 0.6.1__ [2025-02-07] added/testing use.bam option to isolateContribution (might not be staying) 
+    * __v 0.6.2__ [2025-07-12] updated to date handling in calcDateRangeStat; added tinytest package testing   
  
 # AQEval version 0.5
 
