### 2025-07-13 19b2ebe
website and guidance update
