### 2025-07-13 #8338 @karlropkins
@kthyng

I think we were OK if not Good on examples. We have both the R on-board documentation and [introduction on the project website](https://karlropkins.github.io/AQEval/articles/AQEval_Intro_Preprint.pdf). Both include worked examples and discussion of their intended use.

In terms of testing installations, we previously had a set of manual notes which I shared with anyone that contacted us about any problems. I had been happy with that, but, in the spirit of the review process, I figured I should revisit that decision... Long story short: there is now a testing note on the website and I updated package to use [tinytest](https://cran.r-project.org/package=tinytest).

So, I think we are OK/Good on that as well now, but happy to look again at either of these, or any other aspect of the package if there are any concerns.
