### 2025-08-20 #8839 @editorialbot
[REVIEW]: AQEval: R code for the analysis of discrete change in Air Quality time-series
**Submitting author:** @karlropkins (<a href="http://orcid.org/0000-0002-0294-6997">Karl Ropkins</a>)
**Repository:** https://github.com/karlropkins/AQEval
**Branch with paper.md** (empty if default branch): main
**Version:** 0.6.7
