### 2025-11-11 #8839 @karlropkins
Hi, @Sbozzolo, @eldemet and @meenakshi-kushwaha,
I know everyone is really busy and I really appreciate everyone's inputs.
Anything I can do to help, any questions, whatever, please let me know,
And thank you all again,
Karl
