### 2025-11-29 #8839 @karlropkins
@meenakshi-kushwaha Hi,
I've updated the paper; hopefully got everything.
Re 1-2. I've added a Data Requirements and Main Functions section and also linked functions to figures and extended paper, so hopefully that ties up a little better.
Re 3. I've expanded the Summary
It is all a little light-touch, but I am over word count.

@Sbozzolo can you/JOSS live with the word count?
