### 2025-12-07 2423ddb
findNearLatLon
