+++ inst/tinytest/test_AQEval_quant.R
@@ -5,11 +5,14 @@ if(at_home()){
   suppressMessages(ans <- quantBreakPoints(temp, "no2", h=0.15, show="nothing"))
   expect_equal(names(ans), c("data", "breaks", "data2", "plot", "report", "model"))
   expect_equal(ans$breaks$bpt[1], 211)
-  expect_equal(class(ans$plot), c("gg", "ggplot"))
+  # ggplot S3/4/7 ?
+  #expect_equal(class(ans$plot), c("gg", "ggplot"))
+  expect_true(is_ggplot(ans$plot))
 
   suppressMessages(ans <- quantBreakSegments(temp, "no2", h=0.15, show="nothing"))
   expect_equal(names(ans), c("data", "segments", "data2", "plot", "report", "model"))
   expect_equal(ans$segments$seg.str[2], 209)
-  expect_equal(class(ans$plot), c("gg", "ggplot"))
+  #expect_equal(class(ans$plot), c("gg", "ggplot"))
+  expect_true(is_ggplot(ans$plot))
 
 }
