### 2026-01-24 #8839 @eldemet
Thanks @karlropkins for resolving these issues, the paper and software is of high quality.

@Sbozzolo from my side all pending issues have been clarified and I propose the revised paper to be accepted for publicaiton at JOSS. Let me know I need to do anything else at this point.
