### 2026-01-25 #8839 @Sbozzolo
Thanks you, @meenakshi-kushwaha and @eldemet for your review!

@karlropkins, I went over the paper and the repo and left some minor comments below. The submission is in great shape.

1. Your website reads:
> This package contain beta-testing code being developed as part of an on-going project.

Is it still the case (mostly looking at the "beta-testing" part)?

2. For your NEWS file: https://github.com/karlropkins/AQEval/blob/HEAD/NEWS.md

I would recommend sorting your minor versions with newest at the top (the same way you sort the major versions).

3. Readme

Your readme has a truncated sentence (Contributions are very welcome. Please note both contribution guidance and code of conduct… [Contributions](https://karlropkins.github.io/AQEval/CONTRIBUTING.html)). Maybe you want to clean this up?

4. Paper

>  The main functions, quantBreakPoints and quantBreakSegments, use a three-step method to find possible ‘points-of-change’, test these and quantify the most likely or 'regions-of-change' about them, respectively.

I find this sentence quite hard to parse and understand (e.g. what does 'respectively' refer to?). Would you be able to rephrase it?

> Other key functions build of these to provide a workflow

Did you mean "build on these"?

> are unlike to be able to

Did you mean "are unlikely to be able to"?

> Breaks-points are determined using the strucchange methods of Zeileis and colleagues [@zeileis2002strucchange; @zeileis2003testing]. Here, a rolling-window approach is applied: a first subset (time-series window TW0 in \autoref{fig:1}a) is selected and linear regression modelled; the window advanced (TW1 in \autoref{fig:1}a) and a second model built, and so on through the time-series; then, likely points-of-change assigned by comparing the F-Stat scores of sequential models.

This is also a little hard to parse. Is it also missing some "is"?

> In addition to the standard Bayesian Information Criterion (BIC) testing used by strucchange, AQEval also checks all individual break-points are statistically valid (p<0.05), and down-scores less likely combinations.

Is a "that" missing here?

> Where the investigated pollutant concentration

I think the "Where" should be lowercase (as the equation is part of the sentence).

> # to quantify Break-Segments (as in Figure 2 and 3b)
> quantBreakPoints(

Should it be `quantBreakSegments`?

Finally, could you check the DOIs, as reported in the message above?

Thank you!
