+++ joss_paper/paper.bib
@@ -107,7 +107,9 @@
   number={19},
   pages={3055--3071},
   year={2003},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  URL={https://doi.org/10.1002/sim.1545},
+  DOI={10.1002/sim.1545}
 }
 
 @article{muggeo2008segmented,
@@ -159,7 +161,8 @@
   author={R Core Team, R and others},
   year={2025},
   publisher={R foundation for statistical computing Vienna, Austria}, 
-  URL={https://www.R-project.org/}
+  URL={https://doi.org/10.32614/R.manuals},
+  DOI={10.32614/R.manuals}
 }
 
 @article{reeves2007review,
