+++ NEWS.md
@@ -1,42 +1,42 @@
 # AQEval version 0.6
 
 * Package updates 
-    * __v 0.6.0__ [2024-12-19] added data.table method to calcDateRangeStat 
-    * __v 0.6.0__ [2025-01-09] CRAN update 
-    * __v 0.6.1__ [2025-02-07] added/testing use.bam option to isolateContribution (might not be staying) 
-    * __v 0.6.2__ [2025-07-12] updated to date handling in calcDateRangeStat; added tinytest package testing 
-    * __v 0.6.3__ [2025-11-15] find.break typo fixed 
+    * __v 0.6.6__ [2026-01-26] web and doc fixes for JOSS paper 
+    * __v 0.6.5__ [2026-01-15] test update for ggplot class change  
     * __v 0.6.4__ [2025-12-07] findNearLatLon update; option to remove duplicates in ref 
-    * __v 0.6.5__ [2025-01-15] test update for ggplot class change  
+    * __v 0.6.3__ [2025-11-15] find.break typo fixed 
+    * __v 0.6.2__ [2025-07-12] updated to date handling in calcDateRangeStat; added tinytest package testing 
+    * __v 0.6.1__ [2025-02-07] added/testing use.bam option to isolateContribution (might not be staying) 
+    * __v 0.6.0__ [2025-01-09] CRAN update 
+    * __v 0.6.0__ [2024-12-19] added data.table method to calcDateRangeStat 
  
 # AQEval version 0.5
 
 * Package updates 
-    * __v 0.5.0__ [2022-07-20] added (requested) event line and label option for in-call plot 
-    * __v 0.5.1__ [2023-02-23] citation now generated using citation(auto=meta) (requested)
-    * __v 0.5.2__ [2023-02-23] CRAN submission/revision; further citation changes requested for CRAN 
-    * __v 0.5.3__ [2023-07-10] added (requested) extra output options for isolate.contribution 
-    * __v 0.5.4__ [2024-01-09] isolateContribution update; tibble row.name change 
-    * __v 0.5.5__ [2024-01-09] other models; tibble row.name change 
-    * __v 0.5.6__ [2024-04-03] isolateContribution update; week.day fix, year.week added
     * __v 0.5.7__ [2024-04-03] CRAN submission/revision
+    * __v 0.5.6__ [2024-04-03] isolateContribution update; week.day fix, year.week added
+    * __v 0.5.5__ [2024-01-09] other models; tibble row.name change 
+    * __v 0.5.4__ [2024-01-09] isolateContribution update; tibble row.name change 
+    * __v 0.5.3__ [2023-07-10] added (requested) extra output options for isolate.contribution 
+    * __v 0.5.2__ [2023-02-23] CRAN submission/revision; further citation changes requested for CRAN 
+    * __v 0.5.1__ [2023-02-23] citation now generated using citation(auto=meta) (requested)
+    * __v 0.5.0__ [2022-07-20] added (requested) event line and label option for in-call plot 
     
 # AQEval version 0.4
 
 * Package released for CRAN submission 
-    * __v 0.4.0__ [2022-03-31] CRAN submission version
-    * __v 0.4.1__ [2022-04-07] CRAN submission revision; deactivated links not passing win checks  
-    * __v 0.4.2__ [2022-04-07] CRAN submission revision; dubain issue 
-    * __v 0.4.3__ [2022-04-11] CRAN submission revision; CRAN feedback
-    * __v 0.4.4__ [2022-04-11] CRAN submission revision; CRAN feedback (2)
     * __v 0.4.5__ [2022-04-13] CRAN submission revision; CRAN feedback (3)
[…]
