### 2026-01-26 a6e10cd
JOSS paper update
