### 2026-01-26 f500fb2
joss paper update
