+++ joss_paper/paper.md
@@ -1,0 +1,192 @@
+---
+title: 'AQEval: R code for the analysis of discrete change in Air Quality time-series'
+tags: 
+    - R
+    - environmental research
+    - air quality
+    - time-series analysis 
+    - change detection 
+    - break-point/segment
+authors:
+    - name: Karl Ropkins
+      orcid: 0000-0002-0294-6997
+      affiliation: "1"
+    - name: Anthony Walker
+      orcid: 0000-0003-3979-0989
+      affiliation: "2"
+    - name: Ian Philips
+      orcid: 0000-0003-1938-7842
+      affiliation: "1"
+    - name: Christopher E. Rushton
+      orcid: 0000-0002-0159-5187
+      affiliation: "1"
+    - name: Tony Clark
+      affiliation: "2"
+    - name: James E. Tate
+      orcid: 0000-0003-1646-6852
+      affiliation: "1"
+affiliations:
+    - index: 1
+      name: Institute for Transport Studies, University of Leeds, Leeds, LS2 9JT, UK
+    - index: 2
+      name: Joint Air Quality Unit, Department for Transport & Department for Environment, Food and Rural Affairs, Marsham Street London, SW1P 4DF, UK
+date: 29 November 2025
+bibliography: paper.bib
+---
+
+# Summary
+
+`AQEval` (Air Quality Evaluation) is an `R` package for the routine investigation 
+of discrete changes in air quality time-series. The main functions, `quantBreakPoints` 
+and `quantBreakSegments`, use a three-step method to find possible 'points-of-change', 
+test these and quantify the most likely 'points-of-change' or 'regions' about them.
+Other key functions build on these to provide a workflow to measure smaller changes 
+and/or changes in more complex environments. 
+
+# Statement of Need
+
+Authorities responsible for air quality management are typically required to implement and 
+evaluate air quality interventions they adopt [@bradley2019review]. These interventions 
+are often costly, disruptive and unpopular [@glazener2019transforming], and the inherent 
+variability monitoring data hinders impact assessments 
+[@kelly2011impact; @pearce2011quantifying; @jones2012large; @grange2019using]. Various methods 
+have been developed to investigate discrete changes in a wide range of time-series 
+[see e.g. @reeves2007review; @truong2020selective] and several `R` [@r2025r] packages have 
+been developed for their use, e.g. `bcp` [@erdman2008bcp], `changepoint` [@killick2016changepoint], 
+`segmented` [@muggeo2008segmented], and `strucchange` [@zeileis2002strucchange]. Some have even 
+been applied to air quality time-series (see e.g. @carslaw2006change, @carslaw2007detecting). 
+However, many of those tasked with air quality policy assessment, although highly skilled in a 
+wide range of monitoring activities, are unlikely to be able to dedicate sufficient time and resources 
+to the development of in-house expertise in such specialist analyses. `AQEval` was developed to 
+address this skill gap and published as an open research tool. It aligns the inputs and outputs 
+of a number of statistical methods to provide a one-package option for anyone interested in using 
+`R` to routinely investigate change in air quality data.  
+
+# Sources
+
+`AQEval` is freely available under General Public License (GPL): 
+
+-	The latest (stable) release version of `AQEval` is on the Comprehensive R Archive Network (CRAN) 
+[https://CRAN.R-project.org/package=AQEval](https://CRAN.R-project.org/package=AQEval); 
+-	The developers’ version and code are publicly on GitHub 
+[https://github.com/karlropkins/AQEval](https://github.com/karlropkins/AQEval), where issues or change 
+requests can also be posted; 
+-	The project website is at [https://karlropkins.github.io/AQEval/](https://karlropkins.github.io/AQEval/); and
[…]
