+++ joss_paper/paper.bib
@@ -1,0 +1,265 @@
+@article{carslaw2007detecting,
+  title={Detecting and characterising small changes in urban nitrogen dioxide concentrations},
+  author={Carslaw, David C and Carslaw, Nicola},
+  journal={Atmospheric Environment},
+  volume={41},
+  number={22},
+  pages={4723--4733},
+  year={2007},
+  publisher={Elsevier},
+  URL={https://doi.org/10.1016/j.atmosenv.2007.03.034},
+  DOI={10.1016/j.atmosenv.2007.03.034}
+}
+
+@article{carslaw2012openair,
+  title={Openair—an R package for air quality data analysis},
+  author={Carslaw, David C and Ropkins, Karl},
+  journal={Environmental Modelling \& Software},
+  volume={27},
+  pages={52--61},
+  year={2012},
+  publisher={Elsevier},
+  URL={https://doi.org/10.1016/j.envsoft.2011.09.008},
+  DOI={10.1016/j.envsoft.2011.09.008}
+}
+
+@article{carslaw2006change,
+  title={Change-point detection of gaseous and particulate traffic-related pollutants at a roadside location},
+  author={Carslaw, David C and Ropkins, Karl and Bell, Margaret C},
+  journal={Environmental science \& technology},
+  volume={40},
+  number={22},
+  pages={6912--6918},
+  year={2006},
+  publisher={ACS Publications},
+  URL={https://doi.org/10.1016/j.scitotenv.2018.10.344},
+  DOI={10.1016/j.scitotenv.2018.10.344}
+}
+
+@article{erdman2008bcp,
+  title={bcp: an R package for performing a Bayesian analysis of change point problems},
+  author={Erdman, Chandra and Emerson, John W},
+  journal={Journal of statistical Software},
+  volume={23},
+  pages={1--13},
+  year={2008},
+  URL={http://www.jstatsoft.org/v23/i03/}
+}
+
+@article{glazener2019transforming,
+  title={Transforming our cities: best practices towards clean air and active transportation},
+  author={Glazener, Andrew and Khreis, Haneen},
+  journal={Current environmental health reports},
+  volume={6},
+  pages={22--37},
+  year={2019},
+  publisher={Springer}, 
+  URL={https://doi.org/DOI:10.1007/s40572-019-0228-1},
+  DOI={10.1007/s40572-019-0228-1}
+}
+
+@article{grange2019using,
+  title={Using meteorological normalisation to detect interventions in air quality time series},
+  author={Grange, Stuart K and Carslaw, David C},
+  journal={Science of the Total Environment},
+  volume={653},
+  pages={578--588},
+  year={2019},
+  publisher={Elsevier}, 
+  URL={https://doi.org/10.1016/j.scitotenv.2018.10.344},
+  DOI={10.1016/j.scitotenv.2018.10.344}
+}
+
+@article{jones2012large,
+  title={A large reduction in airborne particle number concentrations at the time of the introduction of “sulphur free” diesel and the London Low Emission Zone},
+  author={Jones, Alan M and Harrison, Roy M and Barratt, Benjamin and Fuller, Gary},
+  journal={Atmospheric Environment},
+  volume={50},
+  pages={129--138},
+  year={2012},
+  publisher={Elsevier}, 
+  URL={https://doi.org/10.1016/j.atmosenv.2011.12.050},
+  DOI={10.1016/j.atmosenv.2011.12.050}
+}
+
+@article{kelly2011impact,
+  title={The impact of the congestion charging scheme on air quality in London. Part 1. Emissions modeling and analysis of air pollution measurements.},
[…]
