### 2026-01-26 3599df2
JOSS paper update
