### 2026-01-26 #8839 @karlropkins
Hi @Sbozzolo ,
Many thanks.
Responses…

1. Re Background: Sorry website background was old/out-of-date. We moved out of beta-testing phasing about 2 years ago! Section was updated in readme but not index for website. My mistake. Now updated to same as readme, ‘This package contains code developed as part of an on-going project.’

2. Re News: Agreed, done

3. Re Readme: Thank you for spotting that! Updated (here and in website index) to: ‘Contributions are very welcome. Please check both the [contribution guidance](https://karlropkins.github.io/AQEval/CONTRIBUTING.html) and [code of conduct](https://karlropkins.github.io/AQEval/CODE_OF_CONDUCT.html) when contributing.’

4.	Re Paper:
a.	Sentence ‘The main functions, quantBreakPoints and quantBreakSegments, use a three-step method to find possible ‘points-of-change’, test these and quantify the most likely or 'regions-of-change' about them, respectively.’  changed to ‘The main functions, quantBreakPoints and quantBreakSegments, use a three-step method to find possible ‘points-of-change’, test these and quantify the most likely ‘points-of-change’ or 'regions' about them.’
b.	Yes/sorry/thanks. Changed ‘of’ to ‘on’
c.	Yes/sorry/thanks. Changed ‘unlike’ to ‘unlikely’
d.	Sentence ‘Here, a rolling-window approach is applied: a first subset (time-series window TW0 in \autoref{fig:1}a) is selected and linear regression modelled; the window advanced (TW1 in \autoref{fig:1}a) and a second model built, and so on through the time-series; then, likely points-of-change assigned by comparing the F-Stat scores of sequential models.’ changed to ‘Here, a rolling-window approach is applied: a first subset (time-series window TW0 in \autoref{fig:1}a) is selected and linear regression modelled; then the window is advanced (TW1 in \autoref{fig:1}a) and a second model built, and so on through the time-series; and, finally, likely points-of-change are assigned by comparing the F-Stat scores of sequential models.’
e.	Sentence update: ‘In addition to the standard Bayesian Information Criterion (BIC) testing used by strucchange, AQEval also checks that all individual break-points are statistically valid (p<0.05), and down-scores less likely combinations.’
f.	Agreed/Thanks: ‘Where’ to ‘where’.
g.	Agreed/Particular large thank for spotting that: ‘quantBreakPoints’ to ‘quantBreakSegments’.

5. Re DOIs: Checked. Skipped DOIs are packages or reports we have included URLs for.
Missed DOIs:
a.	DOI (and URL) added to ‘Estimating regression models with unknown…’ in
b.	Added DOI and updated URL for ‘R: A language and environment for…’ (That one feels particularly remiss on my part, sorry)

We really appreciate all your time, input and advice as part of the review process. I suspect first-timers like me/us are a lot of work, but from our/my side, the time and attention really helped. Added following to acknowledgements ‘, and the help, input and advice of editors and reviewers at [JOSS]( https://joss.theoj.org/).’ Does add to word count but hope it is OK?
