### 2026-01-28 #8839 @Sbozzolo
@karlropkins, thank you for your quick response!

Regarding word count, 1317 is acceptable. However, I would encourage you to have another look to see if there's text that could be condensed/removed. For example, the Sources is a little redundant as most of the information there is available in the sidebar or can be immediately be found on the project website. Similarly, Related Outputs could be condensed and added as Reference.

As for DOIs, software packages might also have a DOI. For example, I checked your first reference by following the link and found it there: [10.18637/jss.v023.i03](https://doi.org/10.18637/jss.v023.i03).

Could you have another look at this?

Thank you very much!
