### 2026-01-28 ba8b3aa
joss paper update
