+++ joss_paper/paper.bib
@@ -43,7 +43,8 @@
   volume={23},
   pages={1--13},
   year={2008},
-  URL={http://www.jstatsoft.org/v23/i03/}
+  URL={http://www.jstatsoft.org/v23/i03/},
+  DOI={10.18637/jss.v023.i03}
 }
 
 @article{glazener2019transforming,
@@ -53,7 +54,7 @@
   volume={6},
   pages={22--37},
   year={2019},
-  publisher={Springer}, 
+  publisher={Springer},
   URL={https://doi.org/DOI:10.1007/s40572-019-0228-1},
   DOI={10.1007/s40572-019-0228-1}
 }
@@ -65,7 +66,7 @@
   volume={653},
   pages={578--588},
   year={2019},
-  publisher={Elsevier}, 
+  publisher={Elsevier},
   URL={https://doi.org/10.1016/j.scitotenv.2018.10.344},
   DOI={10.1016/j.scitotenv.2018.10.344}
 }
@@ -77,7 +78,7 @@
   volume={50},
   pages={129--138},
   year={2012},
-  publisher={Elsevier}, 
+  publisher={Elsevier},
   URL={https://doi.org/10.1016/j.atmosenv.2011.12.050},
   DOI={10.1016/j.atmosenv.2011.12.050}
 }
@@ -88,15 +89,19 @@
   journal={Research Report (Health Effects Institute)},
   number={155},
   pages={5--71},
-  year={2011}
+  year={2011},
+  URL={https://europepmc.org/article/MED/21830496}
 }
 
-@misc{killick2016changepoint,
-  title={changepoint: an R package for changepoint analysis. R package version 2.2. 2},
-  author={Killick, R and Haynes, K and Eckley, IA},
-  year={2016},
-  publisher={Comprehensive R Archive Network. https://CRAN.R-project.org/},
-  URL={https://CRAN.R-project.org/package=changepoint}
+@misc{killick2014changepoint,
+  title={changepoint: an R package for changepoint analysis.},
+  author={Killick, R and Eckley, IA},
+  journal={Journal of statistical software},
+  number={58},
+  pages={1--19},
+  year={2014},
+  URL={https://www.jstatsoft.org/article/view/v058i03/0},
+  DOI={10.18637/jss.v058.i03}
 }
 
 @article{muggeo2003estimating,
@@ -119,8 +124,8 @@
   volume={8},
   number={1},
   pages={20--25},
-  year={2008}, 
-  URL={https://cran.r-project.org/doc/Rnews/}
+  year={2008},
+  URL={https://journal.r-project.org/articles/RN-2008-004/RN-2008-004.pdf}
 }
 
 @article{muggeo2017interval,
@@ -131,7 +136,7 @@
   number={3},
   pages={311--322},
   year={2017},
-  publisher={Wiley Online Library}, 
+  publisher={Wiley Online Library},
   URL={https://doi.org/10.1111/anzs.12200},
   DOI={10.1111/anzs.12200}
 }
@@ -144,7 +149,7 @@
   number={6},
   pages={1328--1336},
   year={2011},
-  publisher={Elsevier}, 
+  publisher={Elsevier},
   URL={https://doi.org/10.1016/j.atmosenv.2010.11.051},
   DOI={10.1016/j.atmosenv.2010.11.051}
 }
@@ -152,7 +157,7 @@
 @article{bradley2019review,
   title={Review of interventions to improve outdoor air quality and public health.},
   author={Bradley, N and Dobney, A and Exley, K and Aldridge, JSES and Craswell, A and Dimitroulopoulou, S and Hodgson, G and Izon-Cooper, L and Mitchem, L and Mitsakou, C and others},
-  year={2019}, 
[…]
