### 2026-01-28 #8839 @karlropkins
Hi @Sbozzolo,

Following your suggestions, I have merged Sources and Related Outputs as single shortened section and gone through DOIs. That takes me down from 1,317 to 1,246 words, but think I need to keep some of this because asked for by reviewers as part of their feedback.

Regrading DOIs:
Been through DOIs. Tracked down some; moved some to paper references where I can, or used the package DOI (where I can’t), but FYI, three still reporting as SKIPs.  Details below:

SKIP DOIs
- No DOI given, and none found for title: bcp: an R package for performing a Bayesian analys...
	This should have DOI; was paper no package; my mistake. Sorry.

**- No DOI given, and none found for title: The impact of the congestion charging scheme on ai...
	URL only; EU report; not sure if they have DOI.**

- No DOI given, and none found for title: changepoint: an R package for changepoint analysis...
	Was ref to package which does not have DOI; changed to paper; more relevant because only referencing not using….

**- No DOI given, and none found for title: Segmented: an R package to fit regression models w...
	Have provided the direct link to the r paper pdf;  can’t get reported DOI to work… sorry**

**- No DOI given, and none found for title: Review of interventions to improve outdoor air qua...
	UK Government Report; told no DOI;**

- No DOI given, and none found for title: Generalized additive models: an introduction with ...
	Text book; DOI from publisher;

- No DOI given, and none found for title: mgcv: Mixed GAM Computation Vehicle with Automatic...
	CRAN DOI added; but FYI only takes you to the URL which was already included…

- No DOI given, and none found for title: strucchange: An R package for testing for structur...
	now paper with DOI…
