### 2026-01-31 #8839 @Sbozzolo
Could you change the `misc` in  `@misc{wood2025generalized,` in your bib to be consistent with the other software entries?
