+++ joss_paper/paper.bib
@@ -93,7 +93,7 @@
   URL={https://europepmc.org/article/MED/21830496}
 }
 
-@misc{killick2014changepoint,
+@article{killick2014changepoint,
   title={changepoint: an R package for changepoint analysis.},
   author={Killick, R and Eckley, IA},
   journal={Journal of statistical software},
@@ -239,7 +239,7 @@
   DOI={10.1201/9781315370279}
 }
 
-@misc{wood2025generalized,
+@article{wood2025generalized,
   title={mgcv: Mixed GAM Computation Vehicle with Automatic Smoothness Estimation. R package version 1.9-3},
   author={Wood, Simon N},
   year={2025},
