+++ NEWS.md
@@ -1,6 +1,7 @@
 # AQEval version 0.6
 
 * Package updates 
+    * __v 0.6.7__ [2026-01-31] JOSS paper release  
     * __v 0.6.6__ [2026-01-26] web and doc fixes for JOSS paper 
     * __v 0.6.5__ [2026-01-15] test update for ggplot class change  
     * __v 0.6.4__ [2025-12-07] findNearLatLon update; option to remove duplicates in ref 
