### 2026-01-31 #8839 @karlropkins
Hi @Sbozzolo

Thank you.  😁

- checked authors, orcids, affliations
- made release version - [v.0.6.7](https://github.com/karlropkins/AQEval/releases/tag/AQEval_v067)
- [zenodo copy](https://zenodo.org/records/18440452) DOI:  [10.5281/zenodo.18440451](https://doi.org/10.5281/zenodo.18440451).
- checked author, license, etc
- `@misc` to `@article`
