### 2026-02-03 #8839 @kthyng
@karlropkins
* We recommend that the title and author list for the metadata in your Zenodo software archive exactly match those in your JOSS paper to create a cohesive archive, though this is not a strict requirement. Do you want to make this change?
* Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file. For example "r" and "bayesian" but please check the rest too.
