### 2026-02-04 #8839 @karlropkins
@kthyng
sorry, quick check:
do you want the Zenodo software archive to have the same name as the paper?
