### 2026-02-04 #8839 @kthyng
> do you want the Zenodo software archive to have the same name as the paper?

Yes that is the ask in the first bullet, though it isn't required.
