+++ joss_paper/paper.bib
@@ -12,7 +12,7 @@
 }
 
 @article{carslaw2012openair,
-  title={Openair—an R package for air quality data analysis},
+  title={{o}penair—an {R} package for air quality data analysis},
   author={Carslaw, David C and Ropkins, Karl},
   journal={Environmental Modelling \& Software},
   volume={27},
