+++ joss_paper/paper.bib
@@ -12,7 +12,7 @@
 }
 
 @article{carslaw2012openair,
-  title={{o}penair—an {R} package for air quality data analysis},
+  title={{o}penair — an {R} package for air quality data analysis},
   author={Carslaw, David C and Ropkins, Karl},
   journal={Environmental Modelling \& Software},
   volume={27},
@@ -37,7 +37,7 @@
 }
 
 @article{erdman2008bcp,
-  title={bcp: an R package for performing a Bayesian analysis of change point problems},
+  title={{bcp}: {A}n {R} package for performing a {B}ayesian analysis of change point problems},
   author={Erdman, Chandra and Emerson, John W},
   journal={Journal of statistical Software},
   volume={23},
@@ -48,7 +48,7 @@
 }
 
 @article{glazener2019transforming,
-  title={Transforming our cities: best practices towards clean air and active transportation},
+  title={Transforming our cities: {B}est practices towards clean air and active transportation},
   author={Glazener, Andrew and Khreis, Haneen},
   journal={Current environmental health reports},
   volume={6},
@@ -84,7 +84,7 @@
 }
 
 @article{kelly2011impact,
-  title={The impact of the congestion charging scheme on air quality in London. Part 1. Emissions modeling and analysis of air pollution measurements.},
+  title={The impact of the congestion charging scheme on air quality in {L}ondon. {P}art 1. {E}missions modeling and analysis of air pollution measurements.},
   author={Kelly, Frank and Anderson, H Ross and Armstrong, Ben and Atkinson, Richard and Barratt, Ben and Beevers, Sean and Derwent, Dick and Green, David and Mudway, Ian and Wilkinson, Paul and others},
   journal={Research Report (Health Effects Institute)},
   number={155},
@@ -94,7 +94,7 @@
 }
 
 @article{killick2014changepoint,
-  title={changepoint: an R package for changepoint analysis.},
+  title={{changepoint}: {A}n {R} package for changepoint analysis.},
   author={Killick, R and Eckley, IA},
   journal={Journal of statistical software},
   number={58},
@@ -118,7 +118,7 @@
 }
 
 @article{muggeo2008segmented,
-  title={Segmented: an R package to fit regression models with broken-line relationships},
+  title={{segmented}: {A}n {R} package to fit regression models with broken-line relationships},
   author={Muggeo, Vito MR and others},
   journal={R news},
   volume={8},
@@ -155,14 +155,14 @@
 }
 
 @article{bradley2019review,
-  title={Review of interventions to improve outdoor air quality and public health.},
+  title={Review of interventions to improve outdoor air quality and public health},
   author={Bradley, N and Dobney, A and Exley, K and Aldridge, JSES and Craswell, A and Dimitroulopoulou, S and Hodgson, G and Izon-Cooper, L and Mitchem, L and Mitsakou, C and others},
   year={2019},
   URL={https://assets.publishing.service.gov.uk/government/uploads/system/uploads/attachment_data/file/937341/Principal_interventions_for_local_authorities-air_quality_public_health.pdf}
 }
 
 @article{r2025r,
-  title={R: A language and environment for statistical computing},
+  title={R: {A} language and environment for statistical computing},
   author={R Core Team, R and others},
   year={2025},
   publisher={R foundation for statistical computing Vienna, Austria},
@@ -183,7 +183,7 @@
 }
 
 @article{ropkins2012openair,
-  title={openair: Data Analysis Tools for the Air Quality Community},
+  title={{openair}: {D}ata Analysis Tools for the Air Quality Community},
   author={Ropkins, Karl and Carslaw, David C},
   year={2012},
   journal={The R Journal},
[…]
