+++ joss_paper/paper.bib
@@ -72,7 +72,7 @@
 }
 
 @article{jones2012large,
-  title={A large reduction in airborne particle number concentrations at the time of the introduction of “sulphur free” diesel and the London Low Emission Zone},
+  title={A large reduction in airborne particle number concentrations at the time of the introduction of “sulphur free” diesel and the {L}ondon {L}ow {E}mission {Z}one},
   author={Jones, Alan M and Harrison, Roy M and Barratt, Benjamin and Fuller, Gary},
   journal={Atmospheric Environment},
   volume={50},
@@ -142,7 +142,7 @@
 }
 
 @article{pearce2011quantifying,
-  title={Quantifying the influence of local meteorology on air quality using generalized additive models},
+  title={Quantifying the influence of local meteorology on air quality using {G}eneralized {A}dditive {M}odels},
   author={Pearce, John L and Beringer, Jason and Nicholls, Neville and Hyndman, Rob J and Tapper, Nigel J},
   journal={Atmospheric Environment},
   volume={45},
