### 2026-02-05 #8839 @karlropkins
> > do you want the Zenodo software archive to have the same name as the paper?
>
> Yes that is the ask in the first bullet, though it isn't required.

Sorry over-thinking/thought it should be title of software...

Re bullets:
- https://zenodo.org/records/18440452 updated,
- forced CAPs in .bib

(also updated my own title format so consistent)
