### 2026-02-13 #8839 @kthyng
I am still not seeing the DOI coming through so I will return to this tomorrow.
