### 2026-02-13 #8839 @karlropkins
just checked [0.21105/joss.08839](https://doi.org/10.21105/joss.08839) and I think we are there...
If right, is that paper accepted/in?
